third_party/blink/web_tests/virtual/pna-workers-disabled/external/wpt/fetch/private-network-access/shared-worker.https.window-expected.txt

This is a testharness.js-based test.
FAIL treat-as-public to local: failed preflight. assert_equals: worker error expected (string) "unknown error" but got (undefined) undefined
FAIL treat-as-public to local: success. assert_equals: worker error expected (undefined) undefined but got (string) "unknown error"
FAIL treat-as-public to private: failed preflight. assert_equals: worker error expected (string) "unknown error" but got (undefined) undefined
FAIL treat-as-public to private: success. assert_equals: worker error expected (undefined) undefined but got (string) "unknown error"
PASS public to public: success.
Harness: the test ran to completion.
